python/requirements.txt
aiohttp==1.1.5
astroid==2.3.3
async-timeout==3.0.1
autopep8==1.5.2
certifi==2020.4.5.1
chardet==3.0.4
docutils==0.16
humanfriendly==8.2
idna==2.9
isort==4.3.21
Kivy==1.11.1
Kivy-Garden==0.1.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==4.7.5
pycodestyle==2.5.0
Pygments==2.6.1
pylint==2.4.4
requests==2.23.0
six==1.14.0
speedtest-cli==2.1.2
typed-ast==1.4.1
urllib3==1.25.9
wrapt==1.11.2
yarl==1.4.2
